requirements.txt
Flask==1.1.2
Flask-Cors==3.0.9
Flask-WTF==0.14.3
fpdf==1.7.2
gast==0.2.2
h5py==2.10.0
hdfs==2.5.8
httplib2==0.11.3
idna==2.8
imageio==2.8.0
importlib-metadata==1.7.0
imutils==0.5.3
install==1.3.3
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==0.11
kiwisolver==1.2.0
librosa==0.4.0
livereload==2.6.3
llvmlite==0.31.0
lxml==4.5.1
Mako==1.1.3
Markdown==3.2.1
MarkupSafe==1.1.1
matplotlib==3.2.1
mock==2.0.0
msgpack==1.0.0
msgpack-numpy==0.4.4.3
networkx==2.4
numba==0.48.0
numpy==1.18.2
oauth2client==3.0.0
oauthlib==3.1.0
opencv-python==4.5.1.48
opt-einsum==3.2.0
pandas==1.0.3
pbr==5.4.5
pdf2image==1.14.0
Pillow==7.1.1
pipenv==2020.8.13
protobuf==3.6.1
psutil==5.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
pydot==1.2.4
PyOpenGL==3.1.1a1
pyOpenSSL==19.1.0
pyparsing==2.4.6
PyPDF2==1.26.0
PyQt5==5.14.2
python-dateutil==2.8.1
PyWavelets==1.1.1
requests==2.22.0
requests-oauthlib==1.3.0
slidingwindow==0.0.13
tabulate==0.8.7
termcolor==1.1.0
threadpoolctl==2.0.0
tqdm==4.42.1
urllib3==1.25.8
virtualenv==20.0.31
virtualenv-clone==0.5.4
Wand==0.6.3
Werkzeug==1.0.1
WTForms==2.3.3
xmltodict==0.12.0
gunicorn==20.0.4

